requirements.txt
apache-beam[gcp]==2.2.0
google-cloud-dataflow==2.2.0
git+git://github.com/uh-joan/pdkit@master
pandas-validator==0.5.0
tsfresh==0.11.0
matplotlib==2.2.2
scipy==1.0.1
